requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
openai
chromadb
youtube-transcript-api
python-dotenv
